Renamed use_case package
