+++ src/use_case/logout/LogoutInputData.java
@@ -1,4 +1,4 @@
-package use_case.matching.logout;
+package use_case.logout;
 
 public class LogoutInputData {
     // Doesn't take any input data from user.
