+++ src/use_case/logout/LogoutOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.matching.logout;
+package use_case.logout;
 
 public interface LogoutOutputBoundary {
     void prepareSuccessView(LogoutOutputData user);
